requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
pandas>=2.0.0
streamlit>=1.28.0
python-dotenv>=1.0.0
requests>=2.31.0
